Move are_repos_too_large to code_helpers.py
