+++ fiasko_bro/code_validator.py
@@ -1,5 +1,4 @@
 from collections import OrderedDict
-import os
 
 from . import validators
 from .repository_info import LocalRepositoryInfo
@@ -8,15 +7,6 @@ from .repository_info import LocalRepositoryInfo
 def validate_repo(path_to_repo, path_to_original_repo=None):
     code_validator = CodeValidator()
     return code_validator.validate(path_to_repo, path_to_original_repo)
-
-
-def are_repos_too_large(path_to_repo, path_to_original_repo=None, max_py_files_count):
-    num_of_py_files = count_py_files(path_to_repo)
-    if num_of_py_files > max_py_files_count:
-        return True
-    if path_to_original_repo:
-        num_of_py_files_original_repo = count_py_files(path_to_original_repo)
-        return num_of_py_files_original_repo > max_py_files_count
 
 
 class CodeValidator:
